requirements.txt
Django==2.2.6
sqlparse==0.3.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangotoolbox==1.8.0
django-mongodb-engine-py3==0.6.0.1
pymongo==3.9.0
pytz==2019.3
uWSGI==2.0.18
flake8==3.7.8